Add preliminary Qt6 support
This fixes the build by replacing obsolete APIs that no longer
exist in Qt6, but it has not been tested pretty much at all yet.

+++ src/libclient/export/ffmpegexporter.cpp
@@ -150,7 +150,7 @@ void FfmpegExporter::writeFrame(const QImage &image, int repeat)
 {
 	qInfo("Writing frame (repeat %d)", repeat);
 	if(!m_writebuffer.isEmpty()) {
-		qWarning("FfmpegExporter: tried to write frame while not yet ready! (%lld of %d of buffered bytes written", m_written, m_writebuffer.length());
+		qWarning("FfmpegExporter: tried to write frame while not yet ready! (%lld of %lld of buffered bytes written", m_written, static_cast<long long>(m_writebuffer.length()));
 		return;
 	}
